chore: moving test datasets into a datasets dir
